+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v8.5.0 (2019-07-11)
+
+**⚠️ If you are using React Native 0.60, please follow our migration guide [here](https://github.com/Instabug/Instabug-React-Native/blob/master/README.md#updating-to-version-85)**
+
+* Support for React Native 0.60
+* Updates native iOS and Android SDKs to version 8.5
+
 ## v8.4.4 (2019-07-08)
 
 * Fixes an issue that causes the sdk to crash when a network request has no headers.